refactor(imports): use direct component paths
