+++ src/components/board-buttons/board-buttons.tsx
@@ -24,10 +24,10 @@ import useCountLinksInReplies from '../../hooks/use-count-links-in-replies';
 import useOptimisticReplyCount from '../../hooks/use-optimistic-reply-count';
 import useIsMobile from '../../hooks/use-is-mobile';
 import useTimeFilter from '../../hooks/use-time-filter';
-import CatalogFilters from '../catalog-filters';
-import CatalogSearch from '../catalog-search';
-import Tooltip from '../tooltip';
-import { ModQueueButton } from '../mod-queue-button';
+import CatalogFilters from '../catalog-filters/catalog-filters';
+import CatalogSearch from '../catalog-search/catalog-search';
+import Tooltip from '../tooltip/tooltip';
+import { ModQueueButton } from '../mod-queue-button/mod-queue-button';
 import { isCommentArchived } from '../../lib/utils/comment-moderation-utils';
 import { getSearchWithTimeFilter, getTimeFilterOptionLabel } from '../../lib/utils/time-filter-utils';
 import { shouldShowCatalogButton } from './catalog-button-utils';
